Moved Trip Date to seperate file
